refactor :: logic
