+++ application-infrastructure/src/main/kotlin/hs/kr/equus/application/domain/school/presentation/WebSchoolAdapter.kt
@@ -1,7 +1,7 @@
 package hs.kr.equus.application.domain.school.presentation
 
 import hs.kr.equus.application.domain.school.QuerySchoolUseCase
-import hs.kr.equus.application.domain.school.usecase.dto.response.QuerySchoolResponse
+import hs.kr.equus.application.domain.school.presentation.dto.response.QuerySchoolResponse
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
@@ -16,6 +16,6 @@ class WebSchoolAdapter(
     fun querySchool (
         @RequestParam(value = "school_name")
         name: String
-    ): String =
+    ): QuerySchoolResponse =
         querySchoolUseCase.querySchool(name)
 }
